# internal/infrastructure/http/apiserver/openapi.yaml
openapi: 3.0.3
info:
  title: Alchemorsel API v3 - Pure Backend
  description: |
    Enterprise-grade recipe management API with AI capabilities - Pure JSON Backend
    
    ## Authentication
    This API uses Bearer token authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-token>
    ```
    
    ## Rate Limiting
    - Public endpoints: 100 requests per minute
    - Authenticated endpoints: 1000 requests per minute
    
    ## Response Format
    All responses are in JSON format with consistent error structure:
    ```json
    {
      "success": true,
      "data": {...},
      "message": "Success message"
    }
    ```
    
    Error responses:
    ```json
    {
      "success": false,
      "error": {
        "code": "ERROR_CODE",
        "message": "Human readable error message",
        "details": {...}
      }
    }
    ```
  version: 3.0.0
  termsOfService: https://alchemorsel.com/terms
  contact:
    name: API Support
    url: https://alchemorsel.com/support
    email: support@alchemorsel.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server
  - url: https://api.alchemorsel.com/api/v1
    description: Production server
  - url: https://staging-api.alchemorsel.com/api/v1
    description: Staging server
paths:
...
  /recipes:
    get:
      tags:
        - Recipes
      summary: List recipes
      description: Get a paginated list of recipes with optional filtering
      operationId: listRecipes
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of recipes per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: search
          in: query
          description: Search recipes by name or ingredients
          required: false
          schema:
            type: string
        - name: category
          in: query
          description: Filter by recipe category
          required: false
          schema:
            type: string
        - name: difficulty
          in: query
          description: Filter by difficulty level
          required: false
          schema:
            type: string
            enum: [easy, medium, hard]
        - name: max_time
          in: query
          description: Filter by maximum cooking time in minutes
          required: false
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Recipes retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeListResponse'
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Recipes
      summary: Create a new recipe
      description: Create a new recipe with ingredients and instructions
      operationId: createRecipe
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRecipeRequest'
      responses:
        '201':
          description: Recipe created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /recipes/{id}:
    get:
      tags:
        - Recipes
      summary: Get recipe by ID
      description: Retrieve a specific recipe by its unique identifier
      operationId: getRecipe
      parameters:
        - name: id
          in: path
          description: Recipe unique identifier
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Recipe retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
        '404':
          description: Recipe not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Recipes
      summary: Update recipe
      description: Update an existing recipe (only by recipe owner)
      operationId: updateRecipe
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: Recipe unique identifier
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRecipeRequest'
      responses:
        '200':
          description: Recipe updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - not recipe owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Recipe not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Recipes
      summary: Delete recipe
      description: Delete a recipe (only by recipe owner)
      operationId: deleteRecipe
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: Recipe unique identifier
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Recipe deleted successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - not recipe owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Recipe not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /recipes/{id}/like:
    post:
      tags:
        - Recipes
      summary: Like/unlike a recipe
      description: Toggle like status for a recipe
      operationId: likeRecipe
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: Recipe unique identifier
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Like status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LikeResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Recipe not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /recipes/{id}/rating:
    post:
      tags:
        - Recipes
      summary: Rate a recipe
      description: Add or update a rating for a recipe
      operationId: rateRecipe
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: Recipe unique identifier
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RatingRequest'
      responses:
        '200':
          description: Rating added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RatingResponse'
        '400':
          description: Invalid rating value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Recipe not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter 'Bearer {token}' to authenticate
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: "VALIDATION_ERROR"
            message:
              type: string
              example: "Invalid request data"
            details:
              type: object
          required:
            - code
            - message
      required:
        - success
        - error
...
    Recipe:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        title:
          type: string
          example: "Classic Chocolate Chip Cookies"
        description:
          type: string
          example: "Delicious homemade chocolate chip cookies"
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/Ingredient'
        instructions:
          type: array
          items:
            type: string
          example: ["Preheat oven to 375°F", "Mix dry ingredients", "Add wet ingredients"]
        prep_time:
          type: integer
          description: Preparation time in minutes
          example: 15
        cook_time:
          type: integer
          description: Cooking time in minutes
          example: 12
        servings:
          type: integer
          example: 24
        difficulty:
          type: string
          enum: [easy, medium, hard]
          example: "easy"
        category:
          type: string
          example: "Desserts"
        tags:
          type: array
          items:
            type: string
          example: ["cookies", "dessert", "chocolate"]
        nutrition:
          $ref: '#/components/schemas/NutritionInfo'
        author:
          $ref: '#/components/schemas/User'
        likes_count:
          type: integer
          example: 42
        rating:
          type: number
          format: float
          minimum: 0
          maximum: 5
          example: 4.5
        rating_count:
          type: integer
          example: 15
        created_at:
          type: string
          format: date-time
          example: "2023-12-01T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2023-12-01T10:00:00Z"
      required:
        - id
        - title
        - description
        - ingredients
        - instructions
        - prep_time
        - cook_time
        - servings
        - difficulty
        - category
        - author
        - likes_count
        - rating
        - rating_count
        - created_at
        - updated_at
...
    CreateRecipeRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 200
          example: "Classic Chocolate Chip Cookies"
        description:
          type: string
          minLength: 10
          maxLength: 1000
          example: "Delicious homemade chocolate chip cookies"
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/Ingredient'
          minItems: 1
        instructions:
          type: array
          items:
            type: string
          minItems: 1
          example: ["Preheat oven to 375°F", "Mix dry ingredients", "Add wet ingredients"]
        prep_time:
          type: integer
          minimum: 1
          example: 15
        cook_time:
          type: integer
          minimum: 1
          example: 12
        servings:
          type: integer
          minimum: 1
          example: 24
        difficulty:
          type: string
          enum: [easy, medium, hard]
          example: "easy"
        category:
          type: string
          example: "Desserts"
        tags:
          type: array
          items:
            type: string
          example: ["cookies", "dessert", "chocolate"]
        nutrition:
          $ref: '#/components/schemas/NutritionInfo'
      required:
        - title
        - description
        - ingredients
        - instructions
        - prep_time
        - cook_time
        - servings
        - difficulty
        - category
...
    RecipeListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            recipes:
              type: array
              items:
                $ref: '#/components/schemas/Recipe'
            pagination:
              $ref: '#/components/schemas/Pagination'
          required:
            - recipes
            - pagination
        message:
          type: string
          example: "Recipes retrieved successfully"
      required:
        - success
        - data
        - message
...
    LikeResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            liked:
              type: boolean
              example: true
            likes_count:
              type: integer
              example: 43
          required:
            - liked
            - likes_count
        message:
          type: string
          example: "Recipe liked successfully"
      required:
        - success
        - data
        - message
...